TST: move tests around
